feat: refactoring watch logic for e2e tests
